requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-jazzmin==2.6.0
django-rest-auth==0.9.5
django-rest-authemail==2.1.4
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
pyotp==2.8.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
django-filter~=23.2